# perf(cache): compile ts to js in parallel for `deno cache`