```
Benchmark 1: ../deno/deno_old cache main.ts
  Time (mean ± σ):     225.6 ms ±   6.5 ms    [User: 134.0 ms, System: 80.6 ms]
  Range (min … max):   219.2 ms … 237.2 ms    12 runs
 
Benchmark 2: ../deno/deno_new cache main.ts
  Time (mean ± σ):     168.7 ms ±   6.9 ms    [User: 234.3 ms, System: 167.2 ms]
  Range (min … max):   156.5 ms … 182.0 ms    16 runs
 
Summary
  ../deno/deno_new cache main.ts ran
    1.34 ± 0.07 times faster than ../deno/deno_old cache main.ts
% cat main.ts
import "https://deno.land/x/dax@0.22.0/mod.ts";
import "https://deno.land/x/dax@0.23.0/mod.ts";
import "https://deno.land/x/dax@0.24.0/mod.ts";
import "https://deno.land/x/dax@0.25.0/mod.ts";
```

Closes https://github.com/denoland/deno/issues/23860